Move config loading into framework dir
