+++ src/__tests__/__fixtures__/javaScriptProject/src/getStepStartStates.js
