+++ src/__tests__/__fixtures__/javaScriptProject/src/instanceConfigFields.json
