+++ src/__tests__/__fixtures__/javaScriptProject/src/steps/fetchAccounts.js
