+++ src/__tests__/__fixtures__/typeScriptProject/src/instanceConfigFields.json
