+++ src/__tests__/__fixtures__/typeScriptProject/src/steps/fetchAccounts.ts
