+++ src/__tests__/__fixtures__/typeScriptProject/src/validateInvocation.ts
